agent-driven: Move build_enumerated_prop to ucd_helper.rs
