+++ provider/source/src/properties/enum_codepointtrie.rs
@@ -18,128 +18,7 @@ use std::fmt::Debug;
 use zerotrie::ZeroTrieSimpleAscii;
 use zerovec::ule::NichedOption;
 
-impl SourceDataProvider {
-    #[cfg(any(feature = "use_wasm", feature = "use_icu4c"))]
-    pub(super) fn build_enumerated_prop<T: EnumeratedProperty + Debug>(
-        &self,
-        short_name_to_t: BTreeMap<&'static str, T>,
-    ) -> Result<CodePointTrie<'static, T>, DataError> {
-        let name = core::str::from_utf8(T::NAME).unwrap();
-        let short_name = core::str::from_utf8(T::SHORT_NAME).unwrap();
 
-        self.validate_property_name(name, short_name)?;
-
-        let (names_to_short_names, maybe_default) = self.enumerated_prop_names(name, short_name)?;
-
-        let file = match name {
-            "Indic_Conjunct_Break" => "ucd/DerivedCoreProperties.txt".into(),
-            "Canonical_Combining_Class"
-            | "General_Category"
-            | "Bidi_Class"
-            | "Numeric_Type"
-            | "East_Asian_Width"
-            | "Joining_Type"
-            | "Joining_Group" => {
-                format!(
-                    "ucd/extracted/Derived{}.txt",
-                    name.replace('_', "").replace("Canonical", "")
-                )
-            }
-            "Grapheme_Cluster_Break" | "Word_Break" | "Sentence_Break" => {
-                format!(
-                    "ucd/auxiliary/{}Property.txt",
-                    name.replace('_', "").replace("Cluster", "")
-                )
-            }
-            _ => format!(
-                "ucd/{}.txt",
-                name.replace('_', "").replace("Script", "Scripts")
-            ),
-        };
-
-        let read_to_string = self.unicode()?.read_to_string(&file)?;
-
-        let ucd_default = read_to_string
-            .lines()
-            .find_map(|l| {
-                let mut fields = l
-                    .strip_prefix("# @missing: 0000..10FFFF; ")?
-                    .split(';')
-                    .map(str::trim);
-                if &file == "ucd/DerivedCoreProperties.txt" {
-                    // This is a file containing multiple properties, so we need to check
-                    // the second column for the property name
-                    if fields.next().unwrap() != short_name {
-                        return None;
-                    }
-                }
-                let value = fields.next().unwrap();
-                let value = names_to_short_names
-                    .get(value)
-                    .expect("file should only use names from PropertyValueAliases.txt")
-                    .0;
-                Some(value)
-            })
-            .or_else(|| maybe_default.map(|n| names_to_short_names[n].0))
-            .expect(short_name);
-
-        // @missing entries might use long or short names.
-        let ucd_default = *names_to_short_names
-            .get(ucd_default)
-            .map(|(n, _)| n)
-            .unwrap_or(&ucd_default);
-
-        assert_eq!(
-            *short_name_to_t.get(ucd_default).expect(ucd_default),
-            T::default()
-        );
-
-        let mut builder = icu_codepointtrie_builder::CodePointTrieBuilder::new(
-            T::default(),
-            T::default(),
-            self.trie_type().into(),
-        );
-
-        for line in read_to_string.lines() {
-            let line = line.strip_prefix("# @missing: ").unwrap_or(line);
-            let line = line.split('#').next().unwrap().trim();
-            if line.is_empty() {
-                continue;
-            }
-            let mut fields = line.split(';');
-            let cp_range = fields.next().unwrap().trim();
-            if &file == "ucd/DerivedCoreProperties.txt" {
-                // This is a file containing multiple properties, so we need to check
-                // the second column for the property name
-                if fields.next().unwrap().trim() != short_name {
-                    continue;
-                }
-            }
-            let value = fields.next().unwrap().trim();
-            let value = names_to_short_names
-                .get(value)
-                .expect("file should only use names from PropertyValueAliases.txt")
-                .0;
-            let Some(&value) = short_name_to_t.get(value) else {
-                // Don't log an error for every code point, the name data marker code
-                // will log an error that there's an unknown variant.
-                continue;
-            };
-
-            if let Some((start, end)) = cp_range.split_once("..") {
-                let start = u32::from_str_radix(start, 16).unwrap();
-                let end = u32::from_str_radix(end, 16).unwrap();
-                builder.set_range_value(start..=end, value);
-            } else {
-                let cp = u32::from_str_radix(cp_range, 16).unwrap();
-                builder.set_value(cp, value);
-            }
-        }
-
-        Ok(builder.build())
-    }
-
-}
 
 fn validate_dense<T: TrieValue + Ord + Debug>(map: &BTreeMap<T, &str>) -> Result<(), DataError> {
     if let Some((&first, _)) = map.first_key_value() {
