+++ provider/source/src/properties/ucd_helper.rs
@@ -11,6 +11,7 @@ use icu::properties::PropertyParser;
 use icu::properties::CodePointMapData;
 use icu_provider::prelude::*;
 use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::fmt::Debug;
 use zerovec::VarZeroVec;
 
 /// Type of property name alias.
@@ -330,4 +331,131 @@ impl SourceDataProvider {
 
         Ok((names, default))
     }
+
+    /// Builds a `CodePointTrie` for an enumerated property.
+    ///
+    /// This helper parses the appropriate UCD file, resolves default values (including
+    /// `@missing` rules), and constructs the trie.
+    ///
+    /// Note: This function does NOT use the `parse_ucd_lines` helper for its default-value
+    /// parsing phase because it needs to parse `@missing` metadata which is encoded in UCD comments.
+    #[cfg(any(feature = "use_wasm", feature = "use_icu4c"))]
+    pub(super) fn build_enumerated_prop<T: EnumeratedProperty + Debug>(
+        &self,
+        short_name_to_t: BTreeMap<&'static str, T>,
+    ) -> Result<CodePointTrie<'static, T>, DataError> {
+        let name = core::str::from_utf8(T::NAME).unwrap();
+        let short_name = core::str::from_utf8(T::SHORT_NAME).unwrap();
+
+        self.validate_property_name(name, short_name)?;
+
+        let (names_to_short_names, maybe_default) = self.enumerated_prop_names(name, short_name)?;
+
+        let file = match name {
+            "Indic_Conjunct_Break" => "ucd/DerivedCoreProperties.txt".into(),
+            "Canonical_Combining_Class"
+            | "General_Category"
+            | "Bidi_Class"
+            | "Numeric_Type"
+            | "East_Asian_Width"
+            | "Joining_Type"
+            | "Joining_Group" => {
+                format!(
+                    "ucd/extracted/Derived{}.txt",
+                    name.replace('_', "").replace("Canonical", "")
+                )
+            }
+            "Grapheme_Cluster_Break" | "Word_Break" | "Sentence_Break" => {
+                format!(
+                    "ucd/auxiliary/{}Property.txt",
+                    name.replace('_', "").replace("Cluster", "")
+                )
+            }
+            _ => format!(
+                "ucd/{}.txt",
+                name.replace('_', "").replace("Script", "Scripts")
+            ),
+        };
+
+        let read_to_string = self.unicode()?.read_to_string(&file)?;
+
+        let ucd_default = read_to_string
+            .lines()
+            .find_map(|l| {
+                let mut fields = l
+                    .strip_prefix("# @missing: 0000..10FFFF; ")?
+                    .split(';')
+                    .map(str::trim);
+                if &file == "ucd/DerivedCoreProperties.txt" {
+                    // This is a file containing multiple properties, so we need to check
+                    // the second column for the property name
+                    if fields.next().unwrap() != short_name {
+                        return None;
+                    }
+                }
+                let value = fields.next().unwrap();
+                let value = names_to_short_names
+                    .get(value)
+                    .expect("file should only use names from PropertyValueAliases.txt")
+                    .0;
+                Some(value)
+            })
+            .or_else(|| maybe_default.map(|n| names_to_short_names[n].0))
+            .expect(short_name);
+
+        // @missing entries might use long or short names.
+        let ucd_default = *names_to_short_names
+            .get(ucd_default)
+            .map(|(n, _)| n)
+            .unwrap_or(&ucd_default);
+
+        assert_eq!(
+            *short_name_to_t.get(ucd_default).expect(ucd_default),
+            T::default()
+        );
+
+        let mut builder = icu_codepointtrie_builder::CodePointTrieBuilder::new(
+            T::default(),
+            T::default(),
+            self.trie_type().into(),
+        );
+
+        for line in read_to_string.lines() {
+            let line = line.strip_prefix("# @missing: ").unwrap_or(line);
+            let line = line.split('#').next().unwrap().trim();
+            if line.is_empty() {
+                continue;
+            }
+            let mut fields = line.split(';');
+            let cp_range = fields.next().unwrap().trim();
+            if &file == "ucd/DerivedCoreProperties.txt" {
+                // This is a file containing multiple properties, so we need to check
+                // the second column for the property name
+                if fields.next().unwrap().trim() != short_name {
+                    continue;
+                }
+            }
+            let value = fields.next().unwrap().trim();
+            let value = names_to_short_names
+                .get(value)
+                .expect("file should only use names from PropertyValueAliases.txt")
+                .0;
+            let Some(&value) = short_name_to_t.get(value) else {
+                // Don't log an error for every code point, the name data marker code
+                // will log an error that there's an unknown variant.
+                continue;
+            };
+
+            builder.set_range_value(parse_range(cp_range), value);
+        }
+
+        Ok(builder.build())
+    }
+}
+
+fn parse_range(range_str: &str) -> std::ops::RangeInclusive<u32> {
+    let (a, b) = range_str.split_once("..").unwrap_or((range_str, range_str));
+    let a = u32::from_str_radix(a, 16).unwrap();
+    let b = u32::from_str_radix(b, 16).unwrap();
+    a..=b
 }
